feat: renames open ephys rig
