+++ src/aind_metadata_mapper/open_ephys/rig.py
