+++ tests/test_open_ephys/test_rig.py
@@ -7,7 +7,7 @@ from unittest.mock import MagicMock, patch
 
 from aind_data_schema.core.rig import Rig  # type: ignore
 
-from aind_metadata_mapper.open_ephys.open_ephys_rig import OpenEphysRigEtl
+from aind_metadata_mapper.open_ephys.rig import OpenEphysRigEtl
 
 RESOURCES_DIR = (
     Path(os.path.dirname(os.path.realpath(__file__))) / ".." / "resources"
